+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
     </activity>
 
     <activity
-        android:name="com.maty.android.bookshelves.ui.addBook.AddBookActivity"
+        android:name="com.maty.android.bookshelves.ui.books.add.AddBookActivity"
         android:screenOrientation="portrait"
         android:windowSoftInputMode="stateAlwaysHidden|adjustResize"/>
