--- app/src/main/java/com/maty/android/bookshelves/model/LoginModel.kt
@@ -1,10 +1,0 @@
-package com.maty.android.bookshelves.model
-
-import com.maty.android.bookshelves.common.isEmailValid
-import com.maty.android.bookshelves.common.isPasswordValid
-
-data class LoginRequest(var email: String = "",
-                        var password: String = "") {
-
-  fun isValid() = isEmailValid(email) && isPasswordValid(password)
-}
