+++ app/src/main/java/com/maty/android/bookshelves/ui/books/all/CurrentlyReadingPreviewComponent.kt
@@ -1,4 +1,4 @@
-package com.maty.android.bookshelves.ui.books.all.list
+package com.maty.android.bookshelves.ui.books.all
 
 import android.content.Context
 import android.util.AttributeSet
@@ -11,7 +11,6 @@ import androidx.lifecycle.ViewModelProvider
 import com.maty.android.bookshelves.R
 import com.maty.android.bookshelves.model.Book
 import com.maty.android.bookshelves.ui.books.BookViewModel
-import com.maty.android.bookshelves.ui.books.all.AllBooksView
 import kotlinx.android.synthetic.main.component_booklist_preview.view.noItems
 import kotlinx.android.synthetic.main.component_booklist_preview.view.shimmerLayout
 import kotlinx.android.synthetic.main.component_currently_reading_preview.view.*
