+++ app/src/main/res/layout/fragment_books.xml
@@ -12,27 +12,27 @@
       android:layout_width="match_parent"
       android:layout_height="wrap_content">
 
-    <com.maty.android.bookshelves.ui.books.all.list.CurrentlyReadingPreviewComponent
+    <com.maty.android.bookshelves.ui.books.all.CurrentlyReadingPreviewComponent
         android:id="@+id/booklist_preview_currently_reading"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintTop_toTopOf="parent"/>
 
-    <com.maty.android.bookshelves.ui.books.all.list.BooklistPreviewComponent
+    <com.maty.android.bookshelves.ui.books.all.BooklistPreviewComponent
         android:id="@+id/booklist_preview_to_read"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:booklist_title="@string/books_to_read"
         app:layout_constraintTop_toBottomOf="@id/booklist_preview_currently_reading"/>
 
-    <com.maty.android.bookshelves.ui.books.all.list.BooklistPreviewComponent
+    <com.maty.android.bookshelves.ui.books.all.BooklistPreviewComponent
         android:id="@+id/booklist_preview_to_buy"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:booklist_title="@string/books_to_buy"
         app:layout_constraintTop_toBottomOf="@id/booklist_preview_to_read"/>
 
-    <com.maty.android.bookshelves.ui.books.all.list.BooklistPreviewComponent
+    <com.maty.android.bookshelves.ui.books.all.BooklistPreviewComponent
         android:id="@+id/booklist_preview_already_read"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
